corrected the first task
